fix: installed types for styled-components for twin.macro
