+++ client/src/components/Navbar.tsx
@@ -1,5 +1,4 @@
-import tw from "twin.macro"
-import styled from "styled-components"
+import tw, { styled } from "twin.macro"
 import { ReactComponent as LogoBase } from "../assets/logo.svg"
 import { GradientBar } from "./GradiantBar"
 import { Link } from 'react-router-dom'
@@ -8,8 +7,7 @@ import { myContext } from "../hooks/Context"
 import React, { useContext, useState } from "react";
 import { IUser } from "../interface"
 import { ReactComponent as DiscordIcon } from "../assets/discord-icon.svg";
-import { ReactComponent as GitHubIcon } from "../assets/github-icon.svg";
-import { ReactComponent as TwitterIcon } from "../assets/twitter-icon.svg";
+
 
 const NavContainer = styled.nav`
     ${tw`flex items-center justify-start bg-secondary-700 p-6 lg:p-10 w-full mb-6 lg:mb-12`}
@@ -52,17 +50,8 @@ export default function Navbar() {
 
     const [showModal, setShowModal] = useState(false)
 
-
     const discordLogin = () => {
         window.open("http://localhost:4000/auth/discord?testworking=true&happiness=high", "_self")
-    }
-
-    const githubLogin = () => {
-        window.open("http://localhost:4000/auth/github", "_self")
-    }
-
-    const twitterLogin = () => {
-        window.open("http://localhost:4000/auth/twitter", "_self")
     }
 
     const logout = () => {
@@ -77,7 +66,7 @@ export default function Navbar() {
 
     return (
         <>
-            {showModal ? (
+            { showModal ? (
                 <>
                     <ModalContainer>
                         <ModalContent>
@@ -87,24 +76,16 @@ export default function Navbar() {
                                 </CloseButton>
                             </CloseButtonContainer>
                             <LoginContainer>
-                                <LoginButton onClick={githubLogin}>
-                                    <GitHubIcon />
-                                    Sign in with GitHub
-                                </LoginButton>
                                 <LoginButton onClick={discordLogin}>
                                     <DiscordIcon />
                                     Sign in with Discord
-                                </LoginButton>
-                                <LoginButton onClick={twitterLogin}>
-                                    <TwitterIcon />
-                                    Sign in with Twitter
                                 </LoginButton>
                             </LoginContainer>
                         </ModalContent>
                     </ModalContainer>
                     <BgOpacity />
                 </>
-            ) : null}
+            ) : null }
             <GradientBar />
             <NavContainer>
                 <MainLinks>
@@ -115,7 +96,7 @@ export default function Navbar() {
                     <NavLink to="/featured">Featured</NavLink>
                 </MainLinks>
                 <UserContextLinks>
-                    {user ? (
+                    { user ? (
                         <>
                             <NavLink to="/profile">My Profile</NavLink>
                             <LogoutNavLink onClick={logout}>Logout</LogoutNavLink>
